Audio Files Submitted + 3D Models Submitted (2)
During the development process of the desktop variant to our VR game, I have added a few .obj and .fbx files to test the meshes and collision of custom objects.

Objects implemented:
> Bookcase, Door key, Attack on Titan key and the Clicker. (Excluding the safe mentioned in the previous submission of 3D Models.)
Only the Bookcase, Door key and Safe are textured with proper meshes at  the moment, as .fbx files are difficult to manage.

New audio files submitted (to be used later):
> Door creak sound, Door unlock sound, Optimistic and Sinister UI sounds and Jukebox music.

For note: .obj files are easier to import than .fbx files for 3D objects, and will need some help to implement those properly.